+++ plugin.php
@@ -18,3 +18,4 @@ if ( ! function_exists( __NAMESPACE__ . '\\bootstrap' ) ) {
 add_filter( 'query', __NAMESPACE__ . '\\filter_mysql_query' );
 add_action( 'requests-requests.before_request', __NAMESPACE__ . '\\trace_requests_request', 10, 5 );
 add_filter( 'hm_platform_cloudwatch_error_handler_error', __NAMESPACE__ . '\\on_cloudwatch_error_handler_error' );
+add_filter( 'hm_platform.aws_sdk.params', __NAMESPACE__ . '\\on_hm_platform_aws_sdk_params' );